linux-boundary (3.0.35): Add Vivante GPU 4.6.9p13 support
The 3.10.17-1.0.0 GPU has been updated tp 4.6.9p13 release. This patch
updates the kernel driver to be compatible with it.

Series-to: meta-fsl
Series-prefix: meta-fsl-arm-extra
Series-version: 3
Cover-letter:
Add Vivante 4.6.9p13 GPU support
This adds support for 3.0.35 based boards to work with the new
GPU drivers.

The Wandboard kernel does not build, the GPU patch does not apply and
we need to update it to 3.0.35-4.1.0 or stright to the 3.10.17. I'd
like to get a feedback from John on this.

Please give it a try.
END

Change-Id: Ice6579d97c8488c5e551db5949d3ee3786d50dc9
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>
Tested-by: Eric Nelson <eric.nelson@boundarydevices.com>